requirements.txt
polygon-api-client
pandas
pydantic
pydantic-settings
python-dotenv